pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "black>=25.11.0",
    "isort>=7.0.0",
    "langchain>=1.1.0",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.0",
    "langchain-tavily>=0.2.13",
    "langgraph>=1.0.4",
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
    "tiktoken>=0.12.0",
]
